added attacking animation while ducking
removed the song because the player animation is too big and the rom is going over 48k
tileset tiles were going over 256 individual tiles so increased the tileIndex size in StreamedBatchedAnimationFrame.

+++ project/source/engine/animation_types.h
@@ -99,7 +99,7 @@ typedef struct
 typedef struct streamed_batched_animation_frame
 {
     const BatchedAnimationSprite* const batchedSprites;
-    u8 tileIndex;
+    u16 tileIndex;
     u8 frameTime;
     struct streamed_batched_animation_frame* nextFrame;
 } StreamedBatchedAnimationFrame;
